ManifestInterface now periodically checks for updated manifest
By default, checks once an hour. If you try to set check below 1 minutes, it will be ignored and not check. (so if you dont want it to check, you can set interval to 0 when calling constructor).

Note, you must call init() (which is async) after creating instance of ManifestInterface to initialize class. This is necessary as you cannot have async code in constructors, and I didn't want to have a callback necessary.

```let manifest = new ManifestInterface();
await manifest.init();```