Refactor config navigation
Previously, was using window.location to change app location / player activity overview parameters. This caused the entire page to reload, which wasn't necessary (and kind of defeats the purpose of React). (and yes, this is a dumb issue, but its my first react app).

Now use router.navigate to change pages. However, this was causing an issue, where if you navigated to the same page, with different parameters, the data would not reload. Fixed that in remote.js, by adding the appropriate dependencies.

However, there was still an issue where if you navigated to the page you were on (perhaps to force a data refresh), the data would not up since the parameters were not changing. So now, we pass a flag in the url whose value changes with each submit, that we can use to force the data refresh.

The form to change params is way too intertwined with the date loading in the page. Need to look at something like redux to see if that would be a better solution.